requirements.txt
amazon-kclpy==2.0.6
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
awscli==1.22.49
backcall==0.2.0
black==21.12b0
bleach==4.1.0
boto==2.49.0
boto3==1.20.41
botocore==1.23.49
build==0.7.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
fonttools==4.28.5
idna==3.3
importlib-metadata==4.10.1
ipykernel==6.7.0
ipython==8.0.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==4.4.0
jupyter-client==7.1.1
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keyring==23.5.0
kiwisolver==1.3.2
llvmlite==0.36.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mypy-extensions==0.4.3
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.53.1
numpy==1.20.2
offspring==0.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.2.4
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.9.0
pep517==0.12.0
pickleshare==0.7.5
Pillow==9.0.0
pkginfo==1.8.2
platformdirs==2.4.1
prometheus-client==0.13.1
prompt-toolkit==3.0.24
pure-eval==0.2.1
pyasn1==0.4.8
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.19.2
python-json-logger==2.0.2
pytz==2021.3
pywin32==303
pywin32-ctypes==0.2.0
pywinpty==2.0.1
PyYAML==5.4.1
pyzmq==22.3.0
readme-renderer==32.0
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rsa==4.7.2
s3transfer==0.5.0
Send2Trash==1.8.0
six==1.16.0
stack-data==0.1.4
terminado==0.13.1
testpath==0.5.0
tomli==1.2.3
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
twine==3.7.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.2
zipp==3.7.0
